Added CastAction.cs and relating scripts.
"TierOne" is referring to the amount of steps the raycast takes. For example, while vaulting we use a raycast forward and downwards to confirm our actions affirmation. This would be a "TierTwoRaycast" once this has been scripted.

The is grounded script has been provided as an example, replacing the raycast function within this script with a cast action variable.
Cast Actions are made the same way as other variables, through the asset creation menu. You will then be able to choose the direction and length of the raycast from within the castaction asset.

This system is made to remove having to script out raycasts, during tutorials with SharpAccent. But it can be used wherever you find a place for it.